Refactor after review comments
